+++ deno/payloads/v10/application.ts
@@ -4,6 +4,7 @@
 
 import type { Permissions, Snowflake } from '../../globals.ts';
 import type { LocalizationMap } from '../common.ts';
+import type { APIPartialGuild } from './guild.ts';
 import type { OAuth2Scopes } from './oauth2.ts';
 import type { APITeam } from './teams.ts';
 import type { APIUser } from './user.ts';
@@ -111,6 +112,14 @@ export interface APIApplication {
 	 * which when configured will render the app as a verification method in the guild role verification configuration
 	 */
 	role_connections_verification_url?: string;
+	/**
+	 * An approximate count of the app's guild membership
+	 */
+	approximate_guild_count?: number;
+	/**
+	 * A partial object of the associated guild
+	 */
+	guild?: APIPartialGuild;
 }
 
 export interface APIApplicationInstallParams {