+++ deno/payloads/v10/guild.ts
@@ -3,8 +3,8 @@
  */
 
 import type { Permissions, Snowflake } from '../../globals.ts';
-import type { APIEmoji } from './emoji.ts';
-import type { PresenceUpdateStatus } from './gateway.ts';
+import type { APIEmoji, APIPartialEmoji } from './emoji.ts';
+import type { PresenceUpdateReceiveStatus } from './gateway.ts';
 import type { OAuth2Scopes } from './oauth2.ts';
 import type { APIRole } from './permissions.ts';
 import type { APISticker } from './sticker.ts';
@@ -276,6 +276,10 @@ export interface APIGuild extends APIPartialGuild {
 	 * The type of Student Hub the guild is
 	 */
 	hub_type: GuildHubType | null;
+	/**
+	 * The id of the channel where admins and moderators of Community guilds receive safety alerts from Discord
+	 */
+	safety_alerts_channel_id: Snowflake | null;
 }
 
 /**
@@ -493,6 +497,10 @@ export enum GuildFeature {
 	 * Guild has access to create private threads
 	 */
 	PrivateThreads = 'PRIVATE_THREADS',
+	/**
+	 * Guild has disabled alerts for join raids in the configured safety alerts channel
+	 */
+	RaidAlertsDisabled = 'RAID_ALERTS_DISABLED',
 	RelayEnabled = 'RELAY_ENABLED',
 	/**
 	 * Guild is able to set role icons
@@ -895,7 +903,7 @@ export interface APIGuildWidgetMember {
 	username: string;
 	discriminator: string;
 	avatar: string | null;
-	status: PresenceUpdateStatus;
+	status: PresenceUpdateReceiveStatus;
 	activity?: { name: string };
 	avatar_url: string;
 }
@@ -998,4 +1006,117 @@ export enum MembershipScreeningFieldType {
 	 * Server Rules
 	 */
 	Terms = 'TERMS',
+}
+
+/**
+ * https://discord.com/developers/docs/resources/guild#guild-onboarding-object-guild-onboarding-structure
+ */
+export interface APIGuildOnboarding {
+	/**
+	 * Id of the guild this onboarding is part of
+	 */
+	guild_id: Snowflake;
+	/**
+	 * Prompts shown during onboarding and in customize community
+	 */
+	prompts: APIGuildOnboardingPrompt[];
+	/**
+	 * Channel ids that members get opted into automatically
+	 */
+	default_channel_ids: Snowflake[];
+	/**
+	 * Whether onboarding is enabled in the guild
+	 */
+	enabled: boolean;
+	/**
+	 * Current mode of onboarding
+	 */
+	mode: GuildOnboardingMode;
+}
+
+/**
+ * https://discord.com/developers/docs/resources/guild#guild-onboarding-object-onboarding-prompt-structure
+ */
+export interface APIGuildOnboardingPrompt {
+	/**
+	 * Id of the prompt
+	 */
+	id: Snowflake;
+	/**
+	 * Options available within the prompt
+	 */
+	options: APIGuildOnboardingPromptOption[];
+	/**
+	 * Title of the prompt
+	 */
+	title: string;
+	/**
+	 * Indicates whether users are limited to selecting one option for the prompt
+	 */
+	single_select: boolean;
+	/**
+	 * Indicates whether the prompt is required before a user completes the onboarding flow
+	 */
+	required: boolean;
+	/**
+	 * Indicates whether the prompt is present in the onboarding flow.
+	 * If `false`, the prompt will only appear in the Channels & Roles tab
+	 */
+	in_onboarding: boolean;
+	/**
+	 * Type of prompt
+	 */
+	type: GuildOnboardingPromptType;
+}
+
+/**
+ * https://discord.com/developers/docs/resources/guild#guild-onboarding-object-prompt-option-structure
+ */
+export interface APIGuildOnboardingPromptOption {
+	/**
+	 * Id of the prompt option
+	 */
+	id: Snowflake;
+	/**
+	 * Ids for channels a member is added to when the option is selected
+	 */
+	channel_ids: Snowflake[];
+	/**
+	 * Ids for roles assigned to a member when the option is selected
+	 */
+	role_ids: Snowflake[];
+	/**
+	 * Emoji of the option
+	 */
+	emoji: APIPartialEmoji;
+	/**
+	 * Title of the option
+	 */
+	title: string;
+	/**
+	 * Description of the option
+	 */
+	description: string | null;
+}
+
+/**
+ * https://discord.com/developers/docs/resources/guild#guild-onboarding-object-onboarding-mode
+ */
+export enum GuildOnboardingMode {
+	/**
+	 * Counts only Default Channels towards constraints
+	 */
+	OnboardingDefault,
+	/**
+	 * Counts Default Channels and Questions towards constraints
+	 */
+	OnboardingAdvanced,
+}
+
+/**
+ * https://discord.com/developers/docs/resources/guild#guild-onboarding-object-prompt-types
+ */
+export enum GuildOnboardingPromptType {
+	MultipleChoice,
+	Dropdown,
 }